backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.122.0",
    "google-genai>=1.52.0",
    "google-generativeai>=0.8.5",
    "openai-agents>=0.6.1",
    "pydantic-settings>=2.12.0",
    "qdrant-client>=1.16.1",
]
